Finish ft_printf.c
More testing, and add uintptr_t data type in pointer formatter

+++ cursus/ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: sergio-jimenez <sergio-jimenez@student.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/13 22:04:09 by sergio-jime       #+#    #+#             */
-/*   Updated: 2025/02/17 01:31:22 by sergio-jime      ###   ########.fr       */
+/*   Updated: 2025/02/17 14:17:06 by sergio-jime      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include <stdarg.h>
 # include <unistd.h>
+# include <stdint.h>
 
 int		ft_printf(char const *str, ...);
 int		ft_countchar(char c);
@@ -23,7 +24,7 @@ int		ft_countint(int n);
 int		ft_countunsig(unsigned int n);
 int		ft_counthexalo(unsigned long long n);
 int		ft_counthexaup(unsigned long long n);
-int		ft_countpointer(unsigned long long n);
+int		ft_countpointer(uintptr_t n);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
 
 #endif
